# extension/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAccepted": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
      "NoteRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "library": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Library",
            "type": "string"
          },
          "note_type": {
            "default": "note",
            "enum": [
              "note",
              "question",
              "insight"
            ],
            "title": "Note Type",
            "type": "string"
          },
          "source_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          },
          "timestamp_ms": {
            "minimum": 0.0,
            "title": "Timestamp Ms",
            "type": "integer"
          }
        },
        "required": [
          "library",
          "source_id",
          "timestamp_ms",
          "body"
        ],
        "title": "NoteRequest",
        "type": "object"
      },
      "PairRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 32,
            "minLength": 5,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "PairRequest",
        "type": "object"
      },
      "PairResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at"
        ],
        "title": "PairResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "bili-study Local API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Get Job"
      }
    },
    "/api/v1/libraries": {
      "get": {
        "operationId": "listLibraries",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listlibraries",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Libraries"
      }
    },
    "/api/v1/notes": {
      "post": {
        "operationId": "createNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createnote",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Create Note"
      }
    },
    "/api/v1/pair": {
      "post": {
        "operationId": "pair",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Pair"
      }
    },
    "/api/v1/reflections": {
      "post": {
        "operationId": "createReflection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReflectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Create Reflection"
      }
    },
    "/api/v1/sources/{source_id}/notes": {
      "get": {
        "operationId": "listNotes",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "library",
            "required": true,
            "schema": {
              "title": "Library",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listnotes",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " List Notes"
      }
    },
    "/api/v1/study-guides": {
      "post": {
        "operationId": "createStudyGuide",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Create Study Guide"
      }
    },
    "/api/v1/study-guides/{guide_id}": {
      "get": {
        "operationId": "getStudyGuide",
        "parameters": [
          {
            "in": "path",
            "name": "guide_id",
            "required": true,
            "schema": {
              "title": "Guide Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "library",
            "required": true,
            "schema": {
              "title": "Library",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getstudyguide",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Get Study Guide"
      }
    },
    "/api/v1/study-guides/{guide_id}/chapters/{chapter_id}/details": {
      "post": {
        "operationId": "createChapterDetail",
        "parameters": [
          {
            "in": "path",
            "name": "guide_id",
            "required": true,
            "schema": {
              "title": "Guide Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterDetailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Create Chapter Detail"
      }
    },
    "/api/v1/videos/inspect": {
      "post": {
        "operationId": "inspectVideo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoInspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": " Inspect Video"
      }
    }
  }
}